pyproject.toml
[project]
name = "m165-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.3",
    "flask-cors>=6.0.2",
    "pymongo>=4.16.0",
    "python-dotenv>=1.2.2",
]
